+++ src/main/java/janggi/piece/King.java
@@ -1,7 +1,6 @@
 package janggi.piece;
 
-import janggi.Position;
-import janggi.Side;
+import janggi.board.Position;
 import java.util.List;
 
 public class King extends Piece {
